Replace methods connected with positions from RGBImage to separated PositionManager class [Images/RGBImage.cs, Images/ImageStuff/PositionManager.cs]
